Feature/juliana entities controllers

+++ src/main/java/co/edu/udistrital/mdp/pets/dto/ShelterEventDTO.java
@@ -1,7 +1,6 @@
 package co.edu.udistrital.mdp.pets.dto;
 
 import java.time.LocalDateTime;
-
 import co.edu.udistrital.mdp.pets.entities.ShelterEventEntity;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,7 +10,7 @@ import lombok.NoArgsConstructor;
 public class ShelterEventDTO {
     
     private Long id;
-    private Integer eventCode;
+    private Long eventCode;
     private String eventType;
     private String title;
     private String description;
@@ -21,7 +20,7 @@ public class ShelterEventDTO {
     private int registeredCount;
     private ShelterDTO shelter;
 
-    public ShelterEventDTO (ShelterEventEntity entity) {
+    public ShelterEventDTO(ShelterEventEntity entity) {
         if (entity != null) {
             this.id = entity.getId();
             this.eventCode = entity.getEventCode();
@@ -32,7 +31,7 @@ public class ShelterEventDTO {
             this.location = entity.getLocation();
             this.maxCapacity = entity.getMaxCapacity();
             this.registeredCount = entity.getRegisteredCount();
-
+            
             if (entity.getShelter() != null) {
                 this.shelter = new ShelterDTO(entity.getShelter());
             }
